requirements.txt
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.5
Flask-Session==0.5.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
requests==2.31.0
stripe==5.4.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.2.3
zipp==3.15.0
